requirements.txt
wheel
setuptools
my_utils
pandas
psycopg2-binary